Spring scurity

+++ src/main/java/com/ali/mirsalari/wrench/service/impl/BidServiceImpl.java
@@ -3,11 +3,13 @@ package com.ali.mirsalari.wrench.service.impl;
 import com.ali.mirsalari.wrench.entity.Bid;
 import com.ali.mirsalari.wrench.entity.Expert;
 import com.ali.mirsalari.wrench.entity.Order;
+import com.ali.mirsalari.wrench.entity.User;
 import com.ali.mirsalari.wrench.entity.enumeration.OrderStatus;
 import com.ali.mirsalari.wrench.exception.*;
 import com.ali.mirsalari.wrench.repository.BidRepository;
 import com.ali.mirsalari.wrench.repository.ExpertRepository;
 import com.ali.mirsalari.wrench.repository.OrderRepository;
+import com.ali.mirsalari.wrench.repository.UserRepository;
 import com.ali.mirsalari.wrench.service.BidService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -15,6 +17,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.Instant;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -24,18 +27,19 @@ public class BidServiceImpl implements BidService {
     private final BidRepository bidRepository;
     private final ExpertRepository expertRepository;
     private final OrderRepository orderRepository;
+    private final UserRepository userRepository;
 
     @Override
     public Bid save(Long suggestedPrice, Instant startTime, Instant endTime, Long expertId, Long orderId) {
         Expert expert = expertRepository.findById(expertId).orElseThrow(() -> new NotFoundException("Expert is not found!"));
-        if (expert.getScore()<0){
+        if (expert.getScore() < 0 || !expert.isActive()) {
             throw new DeactivatedAccountException("Account is deactivated due to score less than zero!");
         }
-        Order order = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Order is not found!" ));
+        Order order = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Order is not found!"));
         validation(suggestedPrice, startTime, expert, order);
         Bid bid = new Bid(suggestedPrice, startTime, endTime, expert, order);
         bidRepository.save(bid);
-        if (order.getOrderStatus().equals(OrderStatus.WAITING_FOR_THE_SUGGESTION_OF_EXPERTS)){
+        if (order.getOrderStatus().equals(OrderStatus.WAITING_FOR_THE_SUGGESTION_OF_EXPERTS)) {
             order.setOrderStatus(OrderStatus.WAITING_FOR_EXPERT_SELECTION);
             orderRepository.save(order);
         }
@@ -43,16 +47,17 @@ public class BidServiceImpl implements BidService {
     }
 
     @Override
-    public Bid update(Long id, Long suggestedPrice, Instant startTime, Instant endTime, Long expertId, Long orderId) {
-        Expert expert = expertRepository.findById(expertId).orElseThrow(() -> new NotFoundException("Expert is not found!"));
-        if (expert.getScore()<0){
+    public Bid update(Long id, Long suggestedPrice, Instant startTime, Instant endTime, Long orderId, String email) {
+        Expert expert = expertRepository.findExpertByEmail(email).orElseThrow(() -> new NotFoundException("Expert is not found!"));
+        if (expert.getScore() < 0 || !expert.isActive()) {
             throw new DeactivatedAccountException("Account is deactivated due to score less than zero!");
         }
-        Order order = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Order is not found!" ));
-        validation(suggestedPrice, startTime, expert, order);
-        if (id == null || bidRepository.findById(id).isEmpty()) {
-            throw new NotFoundException("Bid with id: " + id + " is not found.");
+        Bid bid = bidRepository.findById(id).orElseThrow(() -> new NotFoundException("Bid is not found!"));
+        if (!expert.getBids().contains(bid)){
+            throw new IllegalArgumentException("You can not remove this bid!");
         }
+        Order order = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Order is not found!"));
+        validation(suggestedPrice, startTime, expert, order);
         return bidRepository.save(
                 new Bid(
                         id,
@@ -61,17 +66,29 @@ public class BidServiceImpl implements BidService {
                         endTime,
                         expert,
                         order
-        ));
+                ));
     }
 
     @Override
     public Bid updateWithEntity(Bid bid) {
         return bidRepository.save(bid);
     }
+
     @Override
-    public void remove(Long id) {
+    public void remove(Long id, String email) {
+        User user = userRepository.findUserByEmail(email).orElseThrow(() -> new NotFoundException("User is not found!"));
+        if (Objects.equals(user.getUserType(), "ADMIN")) {
+            bidRepository.deleteById(id);
+            return;
+        }
+        Bid bid = bidRepository.findById(id).orElseThrow(() -> new NotFoundException("Bid is not found!"));
+        Expert expert = expertRepository.findExpertByEmail(email).orElseThrow(() -> new NotFoundException("User is not found!"));
+        if (!expert.getBids().contains(bid)) {
+            throw new IllegalStateException("You can not remove this Bid!");
+        }
         bidRepository.deleteById(id);
     }
+
 
     @Override
     public Optional<Bid> findById(Long id) {
@@ -100,14 +117,16 @@ public class BidServiceImpl implements BidService {
         Order order = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Order is not found!"));
         return bidRepository.findRelatedBidsOrderByScoreAsc(order);
     }
+
     @Override
     public List<Bid> findRelatedBidsOrderByScoreDesc(Long orderId) {
         Order order = orderRepository.findById(orderId).orElseThrow(() -> new NotFoundException("Order is not found!"));
         return bidRepository.findRelatedBidsOrderByScoreDesc(order);
     }
+
     @Override
     public void selectBid(Long bidId) {
-        Bid bid = findById(bidId).orElseThrow(()-> new NotFoundException("Bid is not found!"));
+        Bid bid = findById(bidId).orElseThrow(() -> new NotFoundException("Bid is not found!"));
         if (findSelectedBid(bid.getOrder().getId()).isPresent()) {
             throw new DuplicateException("You already have chosen a bid!");
         }
@@ -124,7 +143,7 @@ public class BidServiceImpl implements BidService {
     }
 
     private static void validation(Long suggestedPrice, Instant startTime, Expert expert, Order order) {
-        if (!expert.getSkills().contains(order.getService())){
+        if (!expert.getSkills().contains(order.getService())) {
             throw new IllegalArgumentException("You can not place a bid in this category!");
         }
         if (!(order.getOrderStatus().equals(OrderStatus.WAITING_FOR_THE_SUGGESTION_OF_EXPERTS) ||
