+++ src/main/java/com/ali/mirsalari/wrench/service/impl/CommentServiceImpl.java
@@ -3,16 +3,19 @@ package com.ali.mirsalari.wrench.service.impl;
 import com.ali.mirsalari.wrench.entity.Comment;
 import com.ali.mirsalari.wrench.entity.Customer;
 import com.ali.mirsalari.wrench.entity.Expert;
+import com.ali.mirsalari.wrench.entity.User;
 import com.ali.mirsalari.wrench.exception.NotFoundException;
 import com.ali.mirsalari.wrench.repository.CommentRepository;
 import com.ali.mirsalari.wrench.repository.CustomerRepository;
 import com.ali.mirsalari.wrench.repository.ExpertRepository;
+import com.ali.mirsalari.wrench.repository.UserRepository;
 import com.ali.mirsalari.wrench.service.CommentService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -22,10 +25,11 @@ public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
     private final CustomerRepository customerRepository;
     private final ExpertRepository expertRepository;
+    private final UserRepository userRepository;
 
     @Override
-    public Comment save(byte rate, String verdict, Long customerId, Long expertId) {
-        Customer customer = customerRepository.findById(customerId).orElseThrow(() -> new NotFoundException("Customer is not found!"));
+    public Comment save(byte rate, String verdict, Long expertId, String email) {
+        Customer customer = customerRepository.findCustomerByEmail(email).orElseThrow(() -> new NotFoundException("Customer is not found!"));
         Expert expert = expertRepository.findById(expertId).orElseThrow(() -> new NotFoundException("Expert is not found!"));
         expert.setScore(expert.getScore()+rate);
         expertRepository.save(expert);
@@ -34,16 +38,13 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public Comment update(Long id, byte rate, String verdict, Long customerId, Long expertId) {
+    public Comment update(Long id, byte rate, String verdict, String email) {
         Comment comment = findById(id).orElseThrow(() -> new NotFoundException("Comment is not found!"));
-        Customer customer = customerRepository.findById(customerId).orElseThrow(() -> new NotFoundException("Customer is not found!"));
-        Expert expert = expertRepository.findById(expertId).orElseThrow(() -> new NotFoundException("Expert is not found!"));
-
-        comment.setRate(rate);
-        comment.setVerdict(verdict);
-        comment.setCustomer(customer);
-        comment.setExpert(expert);
-
+        Customer customer = customerRepository.findCustomerByEmail(email).orElseThrow(() -> new NotFoundException("Customer is not found!"));
+        if (comment.getCustomer() == customer) {
+            comment.setRate(rate);
+            comment.setVerdict(verdict);
+        }
         return commentRepository.save(comment);
     }
 
@@ -53,7 +54,17 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void remove(Long id) {
+    public void remove(Long id, String email) {
+        User user = userRepository.findUserByEmail(email).orElseThrow(() -> new NotFoundException("User is not found!"));
+        if (Objects.equals(user.getUserType(), "ADMIN")) {
+            commentRepository.deleteById(id);
+            return;
+        }
+        Customer customer = customerRepository.findCustomerByEmail(email).orElseThrow(()-> new NotFoundException("Customer is not found!"));
+        Comment comment = commentRepository.findById(id).orElseThrow(()-> new NotFoundException("Comment is not found!"));
+        if (comment.getCustomer() != customer){
+            throw new IllegalArgumentException("You can not remove this comment!");
+        }
         commentRepository.deleteById(id);
     }
